fix: corregir concatenación de strings en contador de cartas pokemon

+++ src/app/api/admin/packs/route.ts
@@ -44,8 +44,8 @@ export async function GET(req: NextRequest) {
           s.tcg_id as set_tcg_id,
           s.game,
           s.series,
-          (SELECT COUNT(*) FROM sn_tcg_cards c WHERE c.set_id = p.set_id) as cards_in_set,
-          (SELECT COUNT(*) FROM sn_tcg_cards c WHERE c.set_id = p.set_id AND LOWER(c.name) LIKE LOWER($1)) as pokemon_count
+          (SELECT COUNT(*)::int FROM sn_tcg_cards c WHERE c.set_id = p.set_id) as cards_in_set,
+          (SELECT COUNT(*)::int FROM sn_tcg_cards c WHERE c.set_id = p.set_id AND LOWER(c.name) LIKE LOWER($1)) as pokemon_count
         FROM sn_tcg_packs p
         LEFT JOIN sn_tcg_sets s ON p.set_id = s.id
         WHERE EXISTS (
@@ -72,7 +72,7 @@ export async function GET(req: NextRequest) {
           s.tcg_id as set_tcg_id,
           s.game,
           s.series,
-          (SELECT COUNT(*) FROM sn_tcg_cards c WHERE c.set_id = p.set_id) as cards_in_set
+          (SELECT COUNT(*)::int FROM sn_tcg_cards c WHERE c.set_id = p.set_id) as cards_in_set
         FROM sn_tcg_packs p
         LEFT JOIN sn_tcg_sets s ON p.set_id = s.id
         ORDER BY s.game, s.series, p.name
